+++ envoy-control-runner/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/infrastructure/SynchronizationConfig.kt
@@ -2,12 +2,12 @@ package pl.allegro.tech.servicemesh.envoycontrol.infrastructure
 
 import com.ecwid.consul.v1.ConsulClient
 import io.micrometer.core.instrument.MeterRegistry
-import org.apache.http.impl.client.HttpClients
-import org.apache.http.impl.conn.PoolingHttpClientConnectionManager
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory
+import org.springframework.core.task.SimpleAsyncTaskExecutor
+import org.springframework.http.client.ClientHttpRequestInterceptor
+import org.springframework.http.client.SimpleClientHttpRequestFactory
 import org.springframework.web.client.RestTemplate
 import pl.allegro.tech.discovery.consul.recipes.datacenter.ConsulDatacenterReader
 import pl.allegro.tech.servicemesh.envoycontrol.EnvoyControlProperties
@@ -15,6 +15,7 @@ import pl.allegro.tech.servicemesh.envoycontrol.consul.ConsulProperties
 import pl.allegro.tech.servicemesh.envoycontrol.consul.synchronization.SimpleConsulInstanceFetcher
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.ControlPlaneClient
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.ControlPlaneInstanceFetcher
+import pl.allegro.tech.servicemesh.envoycontrol.synchronization.GzipHttpRequestInterceptor
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.RemoteClusterStateChanges
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.RemoteServices
 import pl.allegro.tech.servicemesh.envoycontrol.synchronization.RestTemplateControlPlaneClient
@@ -26,22 +27,22 @@ import java.util.concurrent.Executors
 class SynchronizationConfig {
 
     @Bean
-    fun restTemplate(envoyControlProperties: EnvoyControlProperties): RestTemplate {
-        val poolingConnManager = PoolingHttpClientConnectionManager()
-            .apply {
-                maxTotal = envoyControlProperties.sync.poolManagerMaxTotal
-                defaultMaxPerRoute = envoyControlProperties.sync.poolManagerDefaultMaxPerRoute
-            }
-
-        val requestFactory = HttpComponentsClientHttpRequestFactory(
-            HttpClients.custom()
-                .setConnectionManager(poolingConnManager)
-                .build()
-        )
+    fun restTemplate(
+        envoyControlProperties: EnvoyControlProperties,
+        interceptors: List<ClientHttpRequestInterceptor>
+    ): RestTemplate {
+        val requestFactory = SimpleClientHttpRequestFactory()
+        requestFactory.setTaskExecutor(SimpleAsyncTaskExecutor())
         requestFactory.setConnectTimeout(envoyControlProperties.sync.connectionTimeout.toMillis().toInt())
         requestFactory.setReadTimeout(envoyControlProperties.sync.readTimeout.toMillis().toInt())
+
         return RestTemplate(requestFactory)
+            .apply { this.interceptors = interceptors }
     }
+
+    @Bean
+    @ConditionalOnProperty(name = ["server.compression.enabled"], havingValue = "true", matchIfMissing = false)
+    fun gzipInterceptor(): GzipHttpRequestInterceptor = GzipHttpRequestInterceptor()
 
     @Bean
     fun controlPlaneClient(restTemplate: RestTemplate, meterRegistry: MeterRegistry, remoteClusters: RemoteClusters) =