+++ envoy-control-runner/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/synchronization/StateController.kt
@@ -1,6 +1,5 @@
 package pl.allegro.tech.servicemesh.envoycontrol.synchronization
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RestController
@@ -9,7 +8,6 @@ import pl.allegro.tech.servicemesh.envoycontrol.services.ServiceInstances
 import pl.allegro.tech.servicemesh.envoycontrol.services.ServicesState
 
 @RestController
-@ConditionalOnProperty(name = ["envoy-control.sync.gzip.enabled"], havingValue = "false", matchIfMissing = true)
 class StateController(val localClusterStateChanges: LocalClusterStateChanges) {
 
     @GetMapping("/state")