+++ CHANGELOG.md
@@ -6,8 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [0.19.33]
 
 ### Changed
-- added switch by flag between default StateController and GzipStateController which produces compressed state response
-- added switch by flag between default and gzip RestTemplateControlPlaneClient 
+- Added compression support for State controller and client
 
 ## [0.19.32]